+++ execution_engine/src/core/engine_state/genesis.rs
@@ -781,7 +781,10 @@ where
         let key = Key::Account(system_account_addr);
         let value = { StoredValue::Account(virtual_system_account.clone()) };
 
-        let _ = tracking_copy.borrow_mut().write(key, value);
+        let _ =
+            tracking_copy
+                .borrow_mut()
+                .write(key, value, executor.config().max_stored_value_size());
 
         GenesisInstaller {
             genesis_config_hash,
@@ -820,6 +823,7 @@ where
                     StoredValue::CLValue(CLValue::from_t(round_seigniorage_rate).map_err(
                         |_| GenesisError::CLValue(ARG_ROUND_SEIGNIORAGE_RATE.to_string()),
                     )?),
+                    self.executor.config().max_stored_value_size(),
                 );
             round_seigniorage_rate_uref
         };
@@ -836,6 +840,7 @@ where
                     CLValue::from_t(U512::zero())
                         .map_err(|_| GenesisError::CLValue(TOTAL_SUPPLY_KEY.to_string()))?,
                 ),
+                self.executor.config().max_stored_value_size(),
             );
             total_supply_uref
         };
@@ -871,6 +876,7 @@ where
             let _ = self.tracking_copy.borrow_mut().write(
                 Key::SystemContractRegistry,
                 StoredValue::CLValue(cl_registry),
+                self.executor.config().max_stored_value_size(),
             );
         }
 
@@ -1034,6 +1040,7 @@ where
                 CLValue::from_t(INITIAL_ERA_ID)
                     .map_err(|_| GenesisError::CLValue(ERA_ID_KEY.to_string()))?,
             ),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(ERA_ID_KEY.into(), era_id_uref.into());
 
@@ -1047,6 +1054,7 @@ where
                 CLValue::from_t(INITIAL_ERA_END_TIMESTAMP_MILLIS)
                     .map_err(|_| GenesisError::CLValue(ERA_END_TIMESTAMP_MILLIS_KEY.to_string()))?,
             ),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(
             ERA_END_TIMESTAMP_MILLIS_KEY.into(),
@@ -1062,6 +1070,7 @@ where
             StoredValue::CLValue(CLValue::from_t(initial_seigniorage_recipients).map_err(
                 |_| GenesisError::CLValue(SEIGNIORAGE_RECIPIENTS_SNAPSHOT_KEY.to_string()),
             )?),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(
             SEIGNIORAGE_RECIPIENTS_SNAPSHOT_KEY.into(),
@@ -1073,6 +1082,7 @@ where
             let _ = self.tracking_copy.borrow_mut().write(
                 Key::Bid(validator_account_hash),
                 StoredValue::Bid(Box::new(bid)),
+                self.executor.config().max_stored_value_size(),
             );
         }
 
@@ -1087,6 +1097,7 @@ where
                 CLValue::from_t(validator_slots)
                     .map_err(|_| GenesisError::CLValue(VALIDATOR_SLOTS_KEY.to_string()))?,
             ),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(VALIDATOR_SLOTS_KEY.into(), validator_slots_uref.into());
 
@@ -1100,6 +1111,7 @@ where
                 CLValue::from_t(auction_delay)
                     .map_err(|_| GenesisError::CLValue(AUCTION_DELAY_KEY.to_string()))?,
             ),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(AUCTION_DELAY_KEY.into(), auction_delay_uref.into());
 
@@ -1113,6 +1125,7 @@ where
                 CLValue::from_t(locked_funds_period_millis)
                     .map_err(|_| GenesisError::CLValue(LOCKED_FUNDS_PERIOD_KEY.to_string()))?,
             ),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(
             LOCKED_FUNDS_PERIOD_KEY.into(),
@@ -1130,6 +1143,7 @@ where
                 CLValue::from_t(unbonding_delay)
                     .map_err(|_| GenesisError::CLValue(UNBONDING_DELAY_KEY.to_string()))?,
             ),
+            self.executor.config().max_stored_value_size(),
         );
         named_keys.insert(UNBONDING_DELAY_KEY.into(), unbonding_delay_uref.into());
 
@@ -1187,7 +1201,11 @@ where
                 main_purse,
             ));
 
-            let _ = self.tracking_copy.borrow_mut().write(key, stored_value);
+            let _ = self.tracking_copy.borrow_mut().write(
+                key,
+                stored_value,
+                self.executor.config().max_stored_value_size(),
+            );
         }
 
         Ok(())
@@ -1328,14 +1346,17 @@ where
         let _ = self.tracking_copy.borrow_mut().write(
             contract_wasm_hash.into(),
             StoredValue::ContractWasm(contract_wasm),
+            self.executor.config().max_stored_value_size(),
         );
-        let _ = self
-            .tracking_copy
-            .borrow_mut()
-            .write(contract_hash.into(), StoredValue::Contract(contract));
+        let _ = self.tracking_copy.borrow_mut().write(
+            contract_hash.into(),
+            StoredValue::Contract(contract),
+            self.executor.config().max_stored_value_size(),
+        );
         let _ = self.tracking_copy.borrow_mut().write(
             contract_package_hash.into(),
             StoredValue::ContractPackage(contract_package),
+            self.executor.config().max_stored_value_size(),
         );
 
         (contract_package_hash, contract_hash)
@@ -1365,6 +1386,7 @@ where
         let _ = self.tracking_copy.borrow_mut().write(
             Key::SystemContractRegistry,
             StoredValue::CLValue(cl_registry),
+            self.executor.config().max_stored_value_size(),
         );
         Ok(())
     }