+++ execution_engine/src/core/engine_state/mod.rs
@@ -28,7 +28,7 @@ use std::{
 use num::Zero;
 use num_rational::Ratio;
 use once_cell::sync::Lazy;
-use tracing::{debug, error};
+use tracing::{debug, error, warn};
 
 use casper_hashing::Digest;
 use casper_types::{
@@ -305,8 +305,11 @@ where
 
         // 3.1.1.1.1.5 bump system contract major versions
         if upgrade_check_result.is_major_version() {
-            let system_upgrader: SystemUpgrader<S> =
-                SystemUpgrader::new(new_protocol_version, tracking_copy.clone());
+            let system_upgrader: SystemUpgrader<S> = SystemUpgrader::new(
+                new_protocol_version,
+                tracking_copy.clone(),
+                self.config.max_stored_value_size(),
+            );
 
             system_upgrader
                 .upgrade_system_contracts_major_version(
@@ -332,7 +335,11 @@ where
                     .map_err(|_| Error::Bytesrepr("new_validator_slots".to_string()))?,
             );
             // Writing a `u32` will not exceed write size limit.
-            let _ = tracking_copy.borrow_mut().write(validator_slots_key, value);
+            let _ = tracking_copy.borrow_mut().write(
+                validator_slots_key,
+                value,
+                self.config.max_stored_value_size(),
+            );
         }
 
         if let Some(new_auction_delay) = upgrade_config.new_auction_delay() {
@@ -346,7 +353,11 @@ where
                     .map_err(|_| Error::Bytesrepr("new_auction_delay".to_string()))?,
             );
             // Writing a `u64` will not exceed write size limit.
-            let _ = tracking_copy.borrow_mut().write(auction_delay_key, value);
+            let _ = tracking_copy.borrow_mut().write(
+                auction_delay_key,
+                value,
+                self.config.max_stored_value_size(),
+            );
         }
 
         if let Some(new_locked_funds_period) = upgrade_config.new_locked_funds_period_millis() {
@@ -360,9 +371,11 @@ where
                     .map_err(|_| Error::Bytesrepr("new_locked_funds_period".to_string()))?,
             );
             // Writing a `u64` will not exceed write size limit.
-            let _ = tracking_copy
-                .borrow_mut()
-                .write(locked_funds_period_key, value);
+            let _ = tracking_copy.borrow_mut().write(
+                locked_funds_period_key,
+                value,
+                self.config.max_stored_value_size(),
+            );
         }
 
         if let Some(new_unbonding_delay) = upgrade_config.new_unbonding_delay() {
@@ -376,7 +389,11 @@ where
                     .map_err(|_| Error::Bytesrepr("new_unbonding_delay".to_string()))?,
             );
             // Writing a `u64` will not exceed write size limit.
-            let _ = tracking_copy.borrow_mut().write(unbonding_delay_key, value);
+            let _ = tracking_copy.borrow_mut().write(
+                unbonding_delay_key,
+                value,
+                self.config.max_stored_value_size(),
+            );
         }
 
         if let Some(new_round_seigniorage_rate) = upgrade_config.new_round_seigniorage_rate() {
@@ -395,15 +412,20 @@ where
                     .map_err(|_| Error::Bytesrepr("new_round_seigniorage_rate".to_string()))?,
             );
             // Writing a `Ratio<U512>` will not exceed write size limit.
-            let _ = tracking_copy
-                .borrow_mut()
-                .write(locked_funds_period_key, value);
+            let _ = tracking_copy.borrow_mut().write(
+                locked_funds_period_key,
+                value,
+                self.config.max_stored_value_size(),
+            );
         }
 
         // apply the arbitrary modifications
         for (key, value) in upgrade_config.global_state_update() {
             // `StoredValue`s produced for the global state update should not exceed write size
             // limit.
+            if value.serialized_length() > self.config.max_stored_value_size() as usize {
+                warn!(%key, serialized_length=%value.serialized_length(), "wrote an upgrade config value which is too large");
+            }
             tracking_copy.borrow_mut().force_write(*key, value.clone());
         }
 
@@ -797,9 +819,11 @@ where
                             mint_extra_keys.push(Key::from(main_purse));
                             // write new account
                             // Writing a default new `Account` will not exceed write size limit.
-                            let _ = tracking_copy
-                                .borrow_mut()
-                                .write(Key::Account(public_key), StoredValue::Account(new_account));
+                            let _ = tracking_copy.borrow_mut().write(
+                                Key::Account(public_key),
+                                StoredValue::Account(new_account),
+                                self.config.max_stored_value_size(),
+                            );
                         }
                         None => {
                             // This case implies that the execution_result is a failure variant as
@@ -1142,6 +1166,7 @@ where
             let _ = tracking_copy.borrow_mut().write(
                 Key::DeployInfo(deploy_item.deploy_hash),
                 StoredValue::DeployInfo(deploy_info),
+                self.config.max_stored_value_size(),
             );
         }
 
@@ -1592,6 +1617,7 @@ where
             let _ = session_tracking_copy.borrow_mut().write(
                 Key::DeployInfo(deploy_hash),
                 StoredValue::DeployInfo(deploy_info),
+                self.config.max_stored_value_size(),
             );
         }