+++ node/src/components/gossiper/tests.rs
@@ -52,6 +52,7 @@ use crate::{
 };
 
 const MAX_ASSOCIATED_KEYS: u32 = 100;
+const MAX_STORED_VALUE_SIZE: u32 = 8 * 1024 * 1024;
 
 /// Top-level event for the reactor.
 #[derive(Debug, From, Serialize)]
@@ -208,6 +209,7 @@ impl reactor::Reactor for Reactor {
             SystemConfig::default(),
             MAX_ASSOCIATED_KEYS,
             DEFAULT_MAX_RUNTIME_CALL_STACK_HEIGHT,
+            MAX_STORED_VALUE_SIZE,
             registry,
         )
         .unwrap();