+++ node/src/types/chainspec/core_config.rs
@@ -32,6 +32,8 @@ pub struct CoreConfig {
     pub(crate) max_associated_keys: u32,
     /// Maximum height of contract runtime call stack.
     pub(crate) max_runtime_call_stack_height: u32,
+    /// Maximum serialized size of values stored in global state.
+    pub(crate) max_stored_value_size: u32,
 }
 
 #[cfg(test)]
@@ -50,6 +52,7 @@ impl CoreConfig {
         );
         let max_associated_keys = rng.gen();
         let max_runtime_call_stack_height = rng.gen();
+        let max_stored_value_size = rng.gen();
 
         CoreConfig {
             era_duration,
@@ -61,6 +64,7 @@ impl CoreConfig {
             round_seigniorage_rate,
             max_associated_keys,
             max_runtime_call_stack_height,
+            max_stored_value_size,
         }
     }
 }
@@ -77,6 +81,7 @@ impl ToBytes for CoreConfig {
         buffer.extend(self.round_seigniorage_rate.to_bytes()?);
         buffer.extend(self.max_associated_keys.to_bytes()?);
         buffer.extend(self.max_runtime_call_stack_height.to_bytes()?);
+        buffer.extend(self.max_stored_value_size.to_bytes()?);
         Ok(buffer)
     }
 
@@ -90,6 +95,7 @@ impl ToBytes for CoreConfig {
             + self.round_seigniorage_rate.serialized_length()
             + self.max_associated_keys.serialized_length()
             + self.max_runtime_call_stack_height.serialized_length()
+            + self.max_stored_value_size.serialized_length()
     }
 }
 
@@ -102,8 +108,9 @@ impl FromBytes for CoreConfig {
         let (locked_funds_period, remainder) = TimeDiff::from_bytes(remainder)?;
         let (unbonding_delay, remainder) = u64::from_bytes(remainder)?;
         let (round_seigniorage_rate, remainder) = Ratio::<u64>::from_bytes(remainder)?;
-        let (max_associated_keys, remainder) = FromBytes::from_bytes(remainder)?;
-        let (max_runtime_call_stack_height, remainder) = FromBytes::from_bytes(remainder)?;
+        let (max_associated_keys, remainder) = u32::from_bytes(remainder)?;
+        let (max_runtime_call_stack_height, remainder) = u32::from_bytes(remainder)?;
+        let (max_stored_value_size, remainder) = u32::from_bytes(remainder)?;
         let config = CoreConfig {
             era_duration,
             minimum_era_height,
@@ -114,6 +121,7 @@ impl FromBytes for CoreConfig {
             round_seigniorage_rate,
             max_associated_keys,
             max_runtime_call_stack_height,
+            max_stored_value_size,
         };
         Ok((config, remainder))
     }